Fix for issue #351

+++ Habitica/src/com/habitrpg/android/habitica/TaskFormActivity.java
@@ -155,7 +155,9 @@ public class TaskFormActivity extends AppCompatActivity implements AdapterView.O
                         .setMessage(getString(R.string.taskform_delete_message)).setPositiveButton(getString(R.string.yes), new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        task.delete();
+                        if (task != null) {
+                            task.delete();
+                        }
 
                         finish();
                         dismissKeyboard();
